refactor musicbrainzalbum (api)
